DTSCCI-683 Refactor ManageContactInformationCallbackHandler in civil-service - fix test
