+++ src/test/java/uk/gov/hmcts/reform/civil/handler/callback/user/ManageContactInformationCallbackHandlerTest.java
@@ -24,6 +24,11 @@ import uk.gov.hmcts.reform.civil.enums.CaseCategory;
 import uk.gov.hmcts.reform.civil.enums.CaseState;
 import uk.gov.hmcts.reform.civil.enums.RespondentResponseType;
 import uk.gov.hmcts.reform.civil.handler.callback.BaseCallbackHandlerTest;
+import uk.gov.hmcts.reform.civil.handler.callback.user.task.managecontactinformation.PrepareEventTask;
+import uk.gov.hmcts.reform.civil.handler.callback.user.task.managecontactinformation.ShowWarningTask;
+import uk.gov.hmcts.reform.civil.handler.callback.user.task.managecontactinformation.SubmitChangesTask;
+import uk.gov.hmcts.reform.civil.handler.callback.user.task.managecontactinformation.ValidateExpertsTask;
+import uk.gov.hmcts.reform.civil.handler.callback.user.task.managecontactinformation.ValidateWitnessesTask;
 import uk.gov.hmcts.reform.civil.helpers.CaseDetailsConverter;
 import uk.gov.hmcts.reform.civil.model.Address;
 import uk.gov.hmcts.reform.civil.model.CaseData;
@@ -104,7 +109,12 @@ import static uk.gov.hmcts.reform.civil.utils.ManageContactInformationUtils.DEFE
     JacksonAutoConfiguration.class,
     CaseDetailsConverter.class,
     PostcodeValidator.class,
-    PartyValidator.class
+    PartyValidator.class,
+    PrepareEventTask.class,
+    ShowWarningTask.class,
+    SubmitChangesTask.class,
+    ValidateExpertsTask.class,
+    ValidateWitnessesTask.class
 })
 @SuppressWarnings("unchecked")
 class ManageContactInformationCallbackHandlerTest extends BaseCallbackHandlerTest {
